2169 cli burndown list link (#2182)
* CLI burndown list changes for Org Management

* few changes in Org Management

* few more changes

* rename org changes

* project commands changes

* refactoring few changes

* project edit command

* self review

* infer project names

* changes for project delete and show cmd

* removed project members list cmd

* remove project members command

* review changes

* few more changes

* more changes

* review changes

* nit changes

---------

+++ cli/cmd/project/delete.go
@@ -11,9 +11,12 @@ import (
 )
 
 func DeleteCmd(cfg *config.Config) *cobra.Command {
+	var name, path string
+	var force bool
+
 	deleteCmd := &cobra.Command{
-		Use:   "delete <project-name>",
-		Args:  cobra.ExactArgs(1),
+		Use:   "delete",
+		Args:  cobra.NoArgs,
 		Short: "Delete",
 		RunE: func(cmd *cobra.Command, args []string) error {
 			client, err := cmdutil.Client(cfg)
@@ -22,17 +25,44 @@ func DeleteCmd(cfg *config.Config) *cobra.Command {
 			}
 			defer client.Close()
 
+			if !cmd.Flags().Changed("project") {
+				name, err = inferProjectName(cmd.Context(), client, cfg.Org, path)
+				if err != nil {
+					return err
+				}
+			}
+
+			if !force {
+				fmt.Printf("Warn: Deleting the project %q will remove all metadata associated with the project\n", name)
+
+				msg := fmt.Sprintf("Enter %q to confirm deletion", name)
+				project, err := cmdutil.InputPrompt(msg, "")
+				if err != nil {
+					return err
+				}
+
+				if project != name {
+					return fmt.Errorf("Entered incorrect name : %s", name)
+				}
+			}
+
 			_, err = client.DeleteProject(context.Background(), &adminv1.DeleteProjectRequest{
 				OrganizationName: cfg.Org,
-				Name:             args[0],
+				Name:             name,
 			})
 			if err != nil {
 				return err
 			}
 
-			cmdutil.SuccessPrinter(fmt.Sprintf("Deleted project: %v\n", args[0]))
+			cmdutil.SuccessPrinter(fmt.Sprintf("Deleted project: %v\n", name))
 			return nil
 		},
 	}
+
+	deleteCmd.Flags().SortFlags = false
+	deleteCmd.Flags().BoolVar(&force, "force", false, "Delete forcefully, skips the confirmation")
+	deleteCmd.Flags().StringVar(&name, "project", "", "Name")
+	deleteCmd.Flags().StringVar(&path, "path", ".", "Project directory")
+
 	return deleteCmd
 }
